add preservation schema and modify all schemas to include top level folder
