+++ ualrdm-directory-validation/tests/preservation/16823602_01_pass-Fangyue_Zhang_d41d8cd98f00b204e9800998ecf8427e/v01/DATA/README.txt
@@ -1,0 +1,35 @@
+---------------------------------------------
+# Precipitation temporal repackaging into fewer, larger storms delayed seasonal timing of peak photosynthesis in a semi-arid grassland
+
+Preferred citation (DataCite format):
+
+  Zhang, Fangyue; Biederman, Joel A.; Pierce, Nathan A.; Potts, Daniel A; Devine, Charles; Hao, Yanbin; et al. (2021).
+  Precipitation temporal repackaging into fewer, larger storms delayed seasonal timing of peak photosynthesis in a semi-arid grassland.
+  University of Arizona Research Data Repository.
+  Dataset. https://doi.org/10.25422/azu.data.16823602
+
+
+Corresponding Author:
+  Fangyue Zhang, School of Natural Resources and the Environment, fangyuezhang@email.arizona.edu
+
+
+License:
+  CC BY 4.0
+
+
+DOI:
+  https://doi.org/10.25422/azu.data.16823602
+
+
+
+---------------------------------------------
+## Summary
+
+This dataset supports the associated publication "Precipitation temporal repackaging into fewer, larger storms delayed seasonal timing of peak photosynthesis in a semi-arid grassland" https://doi.org/10.1111/1365-2435.13980. 
+
+It includes 1) weekly plot-level flux, 2) _Digitaria californica_ reproductive culm, 3) community structure for five blocks (H1-H5) and four precipitation treatments (S1-S4). Each block consists of a rainout shelter (itself consisting of four randomized plots). Precipitation treatments consist of
+- 3.5 days dry interval (S1)
+- 7 days dry interval (S2)
+- 14 days dry interval (S3)
+- 21 days dry interval (S4)
+
