--- ualrdm-directory-validation/ualrdm-directory-schema.json
@@ -1,134 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft/2019-09/schema",
-  "$id": "http://github.com/UAL-RE/ReBACH/ualrdm-directory-validation/ualrdm-directory-schema.json",
-  "type": "array",
-  "items": {
-    "type": "object",
-    "properties": {
-      "type": {
-        "pattern": "^directory$"
-      },
-      "name": {
-        "pattern": "^v\\d{2}$"
-      },
-      "contents": {
-        "type": "array",
-        "contains": {
-          "anyOf": [
-            {
-              "type": "object",
-              "properties": {
-                "type": {
-                  "pattern": "^directory$"
-                },
-                "name": {
-                  "pattern": "^METADATA$"
-                },
-                "contents": {
-                  "type": "array",
-                  "contains": {
-                    "anyOf": [
-                      {
-                        "type": "object",
-                        "properties": {
-                          "type": {
-                            "pattern": "^file$"
-                          },
-                          "name": {
-                            "pattern": "^curation_final_\\d+.*\\.json$"
-                          }
-                        }
-                      },
-                      {
-                        "type": "object",
-                        "properties": {
-                          "type": {
-                            "pattern": "^file$"
-                          },
-                          "name": {
-                            "pattern": "^file_list_final_\\d+.*\\.json$"
-                          }
-                        }
-                      }
-                    ]
-                  },
-                  "minContains": 2
-                }
-              }
-            },
-            {
-              "type": "object",
-              "properties": {
-                "type": {
-                  "pattern": "^directory$"
-                },
-                "name": {
-                  "pattern": "^DATA$"
-                },
-                "contents": {
-                  "type": "array",
-                  "minItems": 0
-                }
-              }
-            },
-            {
-              "type": "object",
-              "properties": {
-                "type": {
-                  "pattern": "^directory$"
-                },
-                "name": {
-                  "pattern": "^UAL_RDM$"
-                },
-                "contents": {
-                  "type": "array",
-                  "contains": {
-                    "anyOf": [
-                      {
-                        "type": "object",
-                        "properties": {
-                          "type": {
-                            "pattern": "^file$"
-                          },
-                          "name": {
-                            "pattern": "^Deposit_Agreement.*\\.pdf$"
-                          }
-                        }
-                      },
-                      {
-                        "type": "object",
-                        "properties": {
-                          "type": {
-                            "pattern": "^file$"
-                          },
-                          "name": {
-                            "pattern": "Trello\\.pdf$"
-                          }
-                        }
-                      },
-                      {
-                        "type": "object",
-                        "properties": {
-                          "type": {
-                            "pattern": "^file$"
-                          },
-                          "name": {
-                            "pattern": "^ReDATA-DepositReview.*\\.pdf$"
-                          }
-                        }
-                      }
-                    ]
-                  },
-                  "minContains": 3
-                }
-              }
-            }
-            
-          ]
-        },
-        "minContains": 3,
-        "maxContains": 3
-      }
-    }
-  }
-}
